Fix ~1cm gaps between snapped cabinets after save/reload
Two cabinets snapped exactly flush (KitchenAssemblyScene's
snapToNeighbor computes exact float-precision positions) could end up
up to ~1cm apart after a save/reload: x/z were whole-centimeter
integers end to end (DB column, Eloquent cast, Laravel validation
rule), and the frontend rounded each module's position independently
on save — two neighbors' centers rounding in opposite directions
adds up to a visible gap at their shared edge.

- New migration widens kitchen_modules.x/z to DECIMAL(8,2) on
  Postgres via raw SQL (no doctrine/dbal installed, so no ->change());
  SQLite needs no migration since its type-affinity rules already
  store a non-integer value in an "integer" column as REAL rather
  than truncating it.
- KitchenModule casts x/z as float instead of integer.
- KitchenProjectController's validation now accepts numeric, not just
  integer, for modules.*.x/z (both the create and update rule sets).
- mapKitchenPayload keeps 2-decimal precision on save instead of
  Math.round()-ing to whole centimeters.

+++ backend/app/Models/KitchenModule.php
@@ -26,8 +26,11 @@ class KitchenModule extends Model
         'height'   => 'integer',
         'width'    => 'integer',
         'depth'    => 'integer',
-        'x'        => 'integer',
-        'z'        => 'integer',
+        // Not integer — see the widen_kitchen_modules_position_precision
+        // migration: a module snapped flush against a neighbor needs
+        // sub-centimeter precision to stay flush through a save/reload.
+        'x'        => 'float',
+        'z'        => 'float',
         'rotation' => 'integer',
     ];
